pyproject.toml
[project]
name = "v360-align-prep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "customtkinter>=5.2.2",
    "pillow>=12.1.1",
    "pyinstaller>=6.19.0",
]
